requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
pydantic==2.5.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
stripe==7.8.0
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
